+++ Framework-Headers/Frameworks/PhotosUI._framework/Objc/MFMailComposeViewControllerDelegate-Protocol.h
@@ -4,13 +4,13 @@
 //  Copyright (C) 1997-2019 Steve Nygard.
 //
 
-#import <PhotosUI/NSObject-Protocol.h>
+#import <PhotosUICore/NSObject-Protocol.h>
 
 @class MFMailComposeViewController, NSError;
 
 @protocol MFMailComposeViewControllerDelegate <NSObject>
 
 @optional
-- (void)mailComposeController:(MFMailComposeViewController *)arg1 didFinishWithResult:(int)arg2 error:(NSError *)arg3;
+- (void)mailComposeController:(MFMailComposeViewController *)arg1 didFinishWithResult:(long long)arg2 error:(NSError *)arg3;
 @end
