+++ Framework-Headers/PrivateFrameworks/AppleAccountUI._framework/Objc/AAUIAccountCreationRemoteUI.h
@@ -46,7 +46,7 @@
 - (void)loadURLRequest:(id)arg1;
 - (void)loadURLRequest:(id)arg1 addHeaders:(BOOL)arg2;
 - (void)loader:(id)arg1 didFailWithError:(id)arg2;
-- (void)loader:(id)arg1 receivedObjectModel:(id)arg2 actionSignal:(int)arg3;
+- (void)loader:(id)arg1 receivedObjectModel:(id)arg2 actionSignal:(unsigned long long)arg3;
 - (void)objectModel:(id)arg1 pressedButton:(id)arg2 attributes:(id)arg3;
 - (void)objectModel:(id)arg1 pressedLink:(id)arg2 httpMethod:(id)arg3;
 - (void)objectModelDidChange:(id)arg1;
