+++ Framework-Headers/PrivateFrameworks/AppleAccountUI._framework/Objc/AAUICredentialRecoveryController.h
@@ -49,7 +49,7 @@
 - (id)init;
 - (id)initWithRemoteUIURL:(id)arg1 modalPresentation:(BOOL)arg2;
 - (void)loader:(id)arg1 didFailWithError:(id)arg2;
-- (void)loader:(id)arg1 receivedObjectModel:(id)arg2 actionSignal:(int)arg3;
+- (void)loader:(id)arg1 receivedObjectModel:(id)arg2 actionSignal:(unsigned long long)arg3;
 - (void)objectModel:(id)arg1 pressedButton:(id)arg2 attributes:(id)arg3;
 - (void)objectModel:(id)arg1 pressedLink:(id)arg2 httpMethod:(id)arg3;
 - (void)objectModelPressedBack:(id)arg1;
