+++ Framework-Headers/PrivateFrameworks/CommunicationsSetupUI._framework/Objc/CNFRegServerWebViewController.h
@@ -59,7 +59,7 @@
 - (void)loadURL:(id)arg1;
 - (void)loadView;
 - (void)loader:(id)arg1 didFailWithError:(id)arg2;
-- (void)loader:(id)arg1 receivedObjectModel:(id)arg2 actionSignal:(int)arg3;
+- (void)loader:(id)arg1 receivedObjectModel:(id)arg2 actionSignal:(unsigned long long)arg3;
 - (id)logName;
 - (void)objectModel:(id)arg1 pressedButton:(id)arg2 attributes:(id)arg3;
 - (void)objectModel:(id)arg1 pressedLink:(id)arg2 httpMethod:(id)arg3;
