+++ Framework-Headers/PrivateFrameworks/MPUFoundation._framework/Objc/NSLayoutConstraint+MPUAutolayout.h
@@ -10,7 +10,7 @@
 + (id)constraintsByAttachingView:(id)arg1 toView:(id)arg2 alongEdges:(unsigned long long)arg3 insets:(struct UIEdgeInsets)arg4;
 + (id)constraintsByAttachingView:(id)arg1 toView:(id)arg2 inCorner:(unsigned long long)arg3 offset:(struct UIOffset)arg4;
 + (id)constraintsByCenteringAndContainingView:(id)arg1 inView:(id)arg2 insets:(struct UIEdgeInsets)arg3;
-+ (id)constraintsByCenteringView:(id)arg1 withView:(id)arg2 alongAxes:(int)arg3 offset:(struct UIOffset)arg4;
++ (id)constraintsByCenteringView:(id)arg1 withView:(id)arg2 alongAxes:(unsigned long long)arg3 offset:(struct UIOffset)arg4;
 + (id)constraintsBySizingView:(id)arg1 toSize:(struct CGSize)arg2;
 @end
